flaskion/requirements.txt
Flask
flask-socketio
flask-cors
eventlet
watchdog
gunicorn
gevent
pycorex
pydbx-hng
psycopg2-binary
marshmallow
dotenv
bcrypt
pyjwt
